src/uwsgi-flask/requirements.txt
flask
uwsgi
mysql-connector
py-bcrypt